computational/requirements.txt
biopython>=1.83
numpy>=1.26
pandas>=2.1
scipy>=1.12
matplotlib>=3.8
seaborn>=0.13
requests>=2.31
scikit-posthocs>=0.9
